Cisto len zmena idcka z longu na int

+++ src/main/java/com/dejvo/Shop/db/mapper/ProductRowMapper.java
@@ -13,7 +13,7 @@ public class ProductRowMapper implements RowMapper<Product> {
         Product product= new Product();
         product.setId(resultSet.getInt("id"));
         product.setDatetime(resultSet.getTimestamp("created_at"));
-        product.setSellerId(resultSet.getLong("seller_id"));
+        product.setSellerId(resultSet.getInt("seller_id"));
         product.setValue(resultSet.getBigDecimal("value"));
         product.setName(resultSet.getString("name"));
         product.setInfo(resultSet.getString("info"));
